requirements.txt
flask
Flask-WTF
flask-login
flask-sqlalchemy
Pillow
Tensorflow
Keras